Move marker_display
