+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/line_list_marker.cpp
@@ -27,8 +27,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "line_list_marker.h"
-#include "marker_selection_handler.h"
+#include "line_list_marker.hpp"
+#include "marker_selection_handler.hpp"
 #include "rviz/default_plugin/marker_display.h"
 #include "rviz/display_context.h"
