+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/line_list_marker.hpp
@@ -30,7 +30,7 @@
 #ifndef RVIZ_LINE_LIST_MARKER_H
 #define RVIZ_LINE_LIST_MARKER_H
 
-#include "marker_base.h"
+#include "marker_base.hpp"
 
 namespace rviz
 {
