+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/marker_base.hpp
