+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/mesh_resource_marker.cpp
@@ -27,9 +27,9 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "mesh_resource_marker.h"
+#include "mesh_resource_marker.hpp"
 
-#include "marker_selection_handler.h"
+#include "marker_selection_handler.hpp"
 #include "rviz/default_plugin/marker_display.h"
 #include "rviz/selection/selection_manager.h"
